+++ src/react.ts
@@ -3,9 +3,7 @@ import mermaid, { type MermaidConfig } from "mermaid";
 
 export interface UseMermaidProps {
     chart: string;
-    config: MermaidConfig & {
-        themeVariables: Record<string, [string, string] | string>;
-    };
+    config: MermaidConfig;
 }
 
 export interface UseMermaidReturn {
@@ -14,6 +12,8 @@ export interface UseMermaidReturn {
     svg?: string;
 }
 
+let initialized = false;
+
 export function useMermaid({ chart, config }: UseMermaidProps): UseMermaidReturn {
     const id = React.useId();
     const [svg, setSvg] = React.useState("");
@@ -21,20 +21,11 @@ export function useMermaid({ chart, config }: UseMermaidProps): UseMermaidReturn
     const [error, setError] = React.useState<Error | undefined>(undefined);
 
     React.useEffect(() => {
-        mermaid.initialize({
-            ...config,
-            themeVariables: {
-                ...Object.fromEntries(
-                    Object.entries(config.themeVariables).map(([key, value]) => {
-                        if (Array.isArray(value)) {
-                            const [light, dark] = value;
-                            return [key, config.darkMode ? dark : light];
-                        }
-                        return [key, value];
-                    })
-                )
-            }
-        });
+        if (!initialized) {
+            mermaid.initialize(config);
+            initialized = true;
+        }
+
         mermaid
             .render(
                 id.replace(/[^a-zA-Z]+/g, ""),
@@ -48,7 +39,7 @@ export function useMermaid({ chart, config }: UseMermaidProps): UseMermaidReturn
                 setError(err);
                 setIsLoading(false);
             });
-    }, [id, chart, config.darkMode]);
+    }, [id, chart]);
 
     return {
         svg,